refactor(commands): sweep the test-pinned sole-caller helpers into their command modules (#233)
With the module tests relocated in #248, move the leftover sole-caller helpers
(together with their now-relocatable lib.rs tests) out of lib.rs into their
command modules. lib.rs drops 2627 -> ~1.8k lines (9.2k at the start of #217).

- commands/tune.rs    <- extract_plugin_entry_toml
- commands/set.rs     <- find_plugin_line_in_toml, update_plugin_config
- commands/remove.rs  <- remove_plugin_from_toml
- commands/sync.rs    <- classify_held_back, matches_rebuild_filter (+ HeldBackPin + the mk_plugin test helper)
- commands/init.rs    <- write_init_lua_snippet + WriteInitResult; loader_init_snippet (pub(crate), see note)
- commands/add.rs     <- the AddTriggerSuggestion cluster (build_add_suggestion / prompt_lazy_decision / decide_add_lazy_apply / resolve_add_lazy_policy / is_interactive_tty / toml_array_preview) + patch_plugin_entry_triggers

Each helper's `#[cfg(test)]` cases moved with it into a `mod tests` in the
command module. Genuinely shared helpers (resolve_plugin_dst, plural,
select_plugin_url, set_plugin_*, parse_*, replace_plugin_entry_with_ai_toml,
prune_*, build_loader_options, ...) stay in lib.rs per the issue.

Notes:
- `loader_init_snippet` is also called by `lib.rs::print_init_lua_hint_if_missing`
  (shared with generate/sync), so it is `pub(crate)` in `commands::init` and
  reached via `crate::commands::loader_init_snippet` from lib.rs rather than
  duplicated.
- Cleaned a stale `// ── ensure_absent ──` section comment left orphaned in
  lib.rs's `mod tests` when #248 moved the ensure_absent tests to merge.rs.

Behaviour-preserving. cargo make check green: 822 tests, plural doctest
preserved, clippy -D warnings --all-targets, fmt, lock-check.

Closes #233. Refs #217.
